Bug fix, truncated kmers larger than 16

+++ src/sdglib/types/KmerTypes.hpp
@@ -197,7 +197,7 @@ struct kmerPos {
         }
     };
 
-    inline bool operator<(uint32_t const &rhs) const { return kmer < rhs; }
+    inline bool operator<(uint64_t const &rhs) const { return kmer < rhs; }
     const bool operator==(const kmerPos &a) const { return std::tie(kmer, contigID, offset) == std::tie(a.kmer, a.contigID, a.offset);}
 };
